jsp추가, Blog 등록수정삭제 기능

+++ Recipe/src/main/java/com/spring/biz/view/blog/BlogController.java
@@ -7,7 +7,10 @@ import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import com.spring.biz.blog.BlogService;
 import com.spring.biz.blog.BlogVO;
@@ -24,11 +27,29 @@ public class BlogController {
 	}
 	
 	@RequestMapping(value = "/insertBlog.do")
-	public String insertBlog(BlogVO bvo) {
+	public String insertBlog(BlogVO bvo, HttpSession sess) {
 		System.out.println("===Controller의 insertBlog() 실행===");
 		System.out.println("bvo : " + bvo);
+		MemberVO mvo = (MemberVO) sess.getAttribute("member");
+		bvo.setWritten(mvo.getNickname());
+		bvo.setId(mvo.getId());
 		int result = blogService.insertBlog(bvo);
-		return "상세블로그화면.jsp";
+		return "blogSingle";
+	}
+	
+	@RequestMapping(value = "/updateBjsp.do", method=RequestMethod.GET)
+	public String updateBjsp(@RequestParam int blogno, Model model, HttpSession sess) {
+		System.out.println("===Controller의 updateBjsp() 실행===");
+		MemberVO mvo = (MemberVO) sess.getAttribute("member");
+		String logid = mvo.getId();
+		String blogid = blogService.idChk(blogno);
+		if(logid.equals(blogid)) {
+			BlogVO bvo = blogService.getBlog(blogno);
+			model.addAttribute("blog", bvo);
+			return "blogUpdate";
+		}else {
+			return "redirect:blogSingle";
+		}
 	}
 	
 	@RequestMapping(value = "/updateBlog.do")
@@ -36,14 +57,21 @@ public class BlogController {
 		System.out.println("===Controller의 updateBlog() 실행===");
 		System.out.println("bvo : " + bvo);
 		int result = blogService.updateBlog(bvo);
-		return "상세블로그화면.jsp";
+		return "blogSingle";
 	}
 	
-	@RequestMapping(value = "/deleteBlog.do")
-	public String deleteBlog(int blogno) {
+	@RequestMapping(value = "/deleteBlog.do", method=RequestMethod.GET)
+	public String deleteBlog(@RequestParam int blogno, HttpSession sess) {
 		System.out.println("===Controller의 deleteBlog() 실행===");
-		blogService.deleteBlog(blogno);
-		return "Blog";
+		MemberVO mvo = (MemberVO) sess.getAttribute("member");
+		String logid = mvo.getId();
+		String blogid = blogService.idChk(blogno);
+		if(logid.equals(blogid)) {
+			blogService.deleteBlog(blogno);
+			return "Blog";
+		}else {
+			return "redirect:blogSingle";
+		}
 	}
 	
 	@RequestMapping(value = "/getBlogList.do")
@@ -64,10 +92,11 @@ public class BlogController {
 		return "myBlog";
 	}
 	
-	@RequestMapping(value = "/getBlog.do")
-	public String getBlog(int blogno) {
+	@RequestMapping(value = "/getBlog.do", method=RequestMethod.GET)
+	public String getBlog(@RequestParam int blogno, Model model) {
 		System.out.println("===Controller의 getBlog() 실행===");
 		BlogVO bvo = blogService.getBlog(blogno);
-		return "상세블로그화면.jsp";
+		model.addAttribute("blog", bvo);
+		return "Blog";
 	}
 }
